chore: consolidate LOADING_COPY and SIGNED_OUT_COPY constants
Moved local constants from `src/pages/Content/script.ts` to `src/constants.ts` to resolve duplication and centralize shared strings. Renamed `SIGNED_OUT_CTA_COPY` to `SIGNED_OUT_COPY` for consistency. Updated imports in Background, Popup, and Content scripts, and fixed vitest mocks to provide `browser.runtime.id`.
