+++ src/pages/Content/script.ts
@@ -8,12 +8,10 @@ import {
 import { v4 as uuidv4 } from 'uuid';
 import browser from 'webextension-polyfill';
 import { getBrowserStorageValue } from '../../storage';
+import { LOADING_COPY, SIGNED_OUT_COPY } from '../../constants';
 
 const EMAIL_INPUT_QUERY_STRING =
   'input[type="email"], input[name="email"], input[id="email"]';
-
-const LOADING_COPY = 'Hide My Email+ — Loading...';
-const SIGNED_OUT_COPY = 'Please sign in to iCloud';
 
 // A unique CSS class prefix is used to guarantee that the style injected
 // by the extension does not interfere with the existing style of
